* small clean ups

+++ AssistanceSDK/app/src/main/java/de/tudarmstadt/informatik/tk/assistance/sdk/model/api/sensing/SensorApiType.java
@@ -277,7 +277,7 @@ public final class SensorApiType {
             return -1;
         }
 
-        return type.intValue();
+        return type;
     }
 
     /**
